implement HTTP DELETE blog feature so that a blog can be deleted only by the user who added the blog

+++ part4/blog-list/controllers/blogs.js
@@ -59,9 +59,25 @@ blogsRouter.post('/', async (request, response, next) => {
   })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if(!request.token || !decodedToken || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
   try {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const blog = await Blog.findById(request.params.id)
+    console.log(blog)
+    if (blog.user.toString() === user.id.toString()){
+      await Blog.findByIdAndDelete(request.params.id)
+      user.blogs = user.blogs.filter(id => id.toString() !== blog.id.toString())
+      await user.save()
+      response.status(204).end()
+    } else { 
+      response.status(403).send({
+        error: 'user does not have the right to delete the resource'
+      }) 
+    }
   } catch (exception){
     next(exception)
   }
